move files to more proper location
